backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
celery==5.5.3
redis==6.4.0
django-celery-beat==2.8.1
PyMySQL==1.1.2
python-dotenv==1.1.1
cffi==1.17.1
Pillow==10.4.0
